rename: SolutionDetailPage 및 컴포넌트 경로 정리
