+++ frontend/src/components/SolutionDetail/SolutionDetailHeader.tsx
