+++ frontend/src/components/SolutionDetail/SolutionSection.styled.ts
