+++ frontend/src/components/SolutionDetail/index.tsx
