+++ frontend/src/pages/SolutionDetailPage/SolutionDetailPage.styled.ts
